Переименовал пакеты. Создал класс MyValues, где лежит URL сайта, почта и пароль для логина
