+++ src/main/java/pom/ForgotPasswordPage.java
@@ -1,4 +1,4 @@
-package POM;
+package pom;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
