+++ src/main/java/values/MyValues.java
@@ -1,0 +1,9 @@
+package values;
+
+public class MyValues {
+    public static final String BASE_URI ="https://stellarburgers.nomoreparties.site/";
+
+    public static final String email = "aposdroid@ya.ru";
+
+    public static final String password = "12341234";
+}
