+++ src/test/java/login/LoginTest.java
@@ -1,9 +1,9 @@
-package LoginTest;
+package login;
 
-import POM.ForgotPasswordPage;
-import POM.LoginPage;
-import POM.MainPage;
-import POM.RegisterPage;
+import pom.ForgotPasswordPage;
+import pom.LoginPage;
+import pom.MainPage;
+import pom.RegisterPage;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Before;
