+++ src/test/java/registration/SuccessfulRegistrationTest.java
@@ -1,8 +1,8 @@
-package RegistrationTest;
+package registration;
 
-import POM.LoginPage;
-import POM.MainPage;
-import POM.RegisterPage;
+import pom.LoginPage;
+import pom.MainPage;
+import pom.RegisterPage;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
